CustomerChurnPrediction/requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo
scikit-learn
dill
pyaml
mlflow
dagshub
fastapi
uvicorn
-e.